fixed migration
